Ошибка подключения хока к Main. WIP
