+++ src/components/app/app.jsx
@@ -4,6 +4,7 @@ import {Switch, Route, BrowserRouter} from "react-router-dom";
 import {connect} from "react-redux";
 import {selectsFilms, selectsPromo} from "../../store/reducers/app-state/selectors";
 import Main from "../main/main";
+import withChangingVisibleFilms from "../../hocs/with-changing-visible-films/with-changing-visible-films";
 import SignIn from "../sign-in/sign-in";
 import MyFilmsList from "../my-films-list/my-films-list";
 import AddReview from "../add-review/add-review";
@@ -19,6 +20,7 @@ import {filmPropStructure} from "../../utils/validator.prop";
 import {isFavoriteType, JumpTo} from "../../utils/constants";
 import {getMatchingFilm, filterFavoriteFilms} from "../../utils/utils";
 
+const MainWrapped = withChangingVisibleFilms(Main);
 const PlayerWrapped = withPlayer(Player);
 
 const App = ({films, promo}) => {
@@ -30,7 +32,7 @@ const App = ({films, promo}) => {
           exact
           path={JumpTo.MAIN}
           render={({history}) => (
-            <Main
+            <MainWrapped
               promo={promo}
               handleButtonPlayClick={(id) => history.push(JumpTo.PLAYER + id)}
             />
